+++ RELEASENOTES.md
@@ -93,6 +93,7 @@
 *   GL demo app:
     *   Fix texture transformation to avoid green bars shown on some videos
         ([#8992](https://github.com/google/ExoPlayer/issues/8992)).
+    *   Add `RtspMediaSource.Factory` option to force using TCP for streaming.
 
 ### 2.14.0 (2021-05-13)